requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
groq
mangum
